adds parse to utils
