+++ src/Map.js
@@ -13,7 +13,7 @@ import { scaleLinear } from "d3-scale";
 import Nav from "./Nav";
 import Filter from "./Filter";
 import Heatmap from "./Heatmap";
-import { getMapIcon } from "./utils";
+import { getMapIcon, parsePrice } from "./utils";
 import useData from "./use-data";
 
 const useStyles = makeStyles(() => ({
@@ -63,35 +63,15 @@ const Map = ({ token, removeToken }) => {
     return scaleLinear().domain([min, max]).range(["red", "steelblue"]);
   }, [distributions]);
 
-  const purchaseGradient = useMemo(() => {
-    let min = 0;
-    let max = 0;
-
-    for (let i = 0; i < purchases.length; i += 1) {
-      const { june } = purchases[i];
-      // Hack to see more stuff.
-      const month = (june || "$0.00")
-        .split("$")[1]
-        .replace(",", "")
-        .replace(".00", "");
-
-      if (month < min) {
-        min = month;
-      }
-
-      if (month > max) {
-        max = month;
-      }
-    }
-
-    return scaleLinear().domain([min, max]).range(["pink", "purple"]);
-  }, [purchases]);
-
   // todo delete
   const filteredDistributions = useMemo(() => distributions, [distributions]);
 
   // todo delete
   const filteredPurchases = useMemo(() => purchases, [purchases]);
+
+  function getPurchaseAmount(purchase) {
+    return 800;
+  }
 
   return (
     <div>
@@ -107,6 +87,7 @@ const Map = ({ token, removeToken }) => {
           isHeatmap,
           showPurchases,
           showDistributions,
+          purchaseGradient,
         }) => (
           <MapContainer center={position} zoom={8} className={classes.root}>
             <TileLayer
@@ -126,18 +107,38 @@ const Map = ({ token, removeToken }) => {
                     style={{ border: 0 }}
                   >
                     <Popup onOpen={() => console.log(item)}>
-                      <strong>Name: </strong>
-                      {item.name}
-                      <br />
-                      <strong>Address: </strong>
-                      {item.address}
-                      <br />
-                      <strong>Category: </strong>
-                      {item.category.join(", ")}
+                      <div style={{ display: "flex" }}>
+                        {item.locationImage && (
+                          <div style={{ width: 120, paddingRight: 30 }}>
+                            <img
+                              style={{ width: "100%" }}
+                              src={item.locationImage}
+                              alt=""
+                            />
+                          </div>
+                        )}
+                        <div>
+                          <strong>Name: </strong>
+                          {item.name}
+                          <br />
+                          {item.description && (
+                            <>
+                              <strong>Description: </strong>
+                              <span>{item.description}</span>
+                              <br />
+                            </>
+                          )}
+                          <strong>Address: </strong>
+                          {item.address}
+                          <br />
+                          <strong>Category: </strong>
+                          {item.category.join(", ")}
+                        </div>
+                      </div>
                     </Popup>
                   </Marker>
                 ))}
-                {showDistributions &&
+                {/* {showDistributions &&
                   filteredDistributions.map(
                     ({ id, hubGeo, distributionSiteGeo, boxes }) => (
                       <Polyline
@@ -146,7 +147,7 @@ const Map = ({ token, removeToken }) => {
                         pathOptions={{ color: distributionGradient(boxes) }}
                       />
                     )
-                  )}
+                  )} */}
                 {showPurchases &&
                   filteredPurchases.map((purchase) => (
                     <Polyline
@@ -155,7 +156,9 @@ const Map = ({ token, removeToken }) => {
                         purchase.hubOrganizationGeo,
                         purchase.farmNameGeo,
                       ]}
-                      pathOptions={{ color: purchaseGradient(800) }}
+                      pathOptions={{
+                        color: purchaseGradient(getPurchaseAmount(purchase)),
+                      }}
                     />
                   ))}
               </>
